Fixup transaction dates with February 30
German banks use a variant of the 30/360 interest rate calculation system,
where each month has exactly 30 days. This can lead to transactions that are
dated on February 30, which are not accepted by many date handling software,
such as Python's datetime module.

A statement pre-processor is added which sets the date of transactions with
an original date of February 30 to the last day of February (February 28 or
February 29 depending on leap years).